static/tests/secure/create-cookie-js.html

<!DOCTYPE html>
<script src="/tests/resources/testharness.js"></script>
<script src="/tests/resources/testharnessreport.js"></script>
<script src="/tests/resources/cookie-helper.js"></script>
<script>
  function create_test(origin_page, params, shouldExistInDOM, shouldExistViaHTTP, title) {
    promise_test(t => {
      var name = "alone_js_testcookie";
      
      //TODO - make sure all deletes happen from a secure origin
      erase_cookie_from_js(name); //CAUTION
      var value = "" + Math.random();
      document.cookie = name + "=" + value + ";" + params;
    
      //TODO - redirect to origin_page?
      assert_dom_cookie(name, value, shouldExistInDOM);

      return credFetch("/cookie/list")
        .then(r => r.json())
        .then(cookies => assert_equals(cookies[name], shouldExistViaHTTP ? value : undefined));
    }, title);
  }

  //
  //Assuming initial origin page is insecure
  //
  
  //basic test
  create_test(INSECURE_ORIGIN, "path=/;", true, true, "insecure origin, root path, no special behavior");
  //You cannot set a Secure cookie via JS
  create_test(INSECURE_ORIGIN, "path=/;Secure;", false, false, "insecure origin, root path, secure cookie");
  //You cannot set a HTTPOnly cookie via JS
  create_test(INSECURE_ORIGIN, "path=/;HttpOnly;", false, false, "insecure origin, root path, httponly cookie");
  
  //
  //Assuming initial origin page is secure
  //
  
  //basic test
  create_test(SECURE_ORIGIN, "path=/;", true, true, "secure origin, root path, no special behavior");
  //You cannot set a Secure cookie via JS
  create_test(SECURE_ORIGIN, "path=/;Secure;", true, true, "secure origin, root path, secure cookie");
  //You cannot set a HTTPOnly cookie via JS
  create_test(SECURE_ORIGIN, "path=/;HttpOnly;", false, true, "secure origin, root path, httponly cookie");
  
 
</script>
